+++ src/components/logos/footer-logo.jsx
