+++ src/index.jsx
@@ -11,7 +11,7 @@ import { config } from 'constants/config';
 import App from 'app';
 import Landing from 'pages/landing';
 import FieldBook from 'pages/field-book';
-import ActiveExpeditions from 'pages/active-expeditions';
+import ActiveExpeditionGroup from 'pages/active-expedition-group';
 import CompletedExpeditions from 'pages/completed-expeditions';
 import CompletedExpeditionGroup from 'pages/completed-expedition-group';
 
@@ -28,7 +28,7 @@ oauth.init(config.panoptesAppId)
         <Router history={browserHistory}>
           <Route path="/" component={App}>
             <IndexRoute component={Landing} />
-            <Route path="/active-expeditions/:group" component={ActiveExpeditions} />
+            <Route path="/active-expeditions/:group" component={ActiveExpeditionGroup} />
             <Route path="/completed-expeditions" component={CompletedExpeditions} />
             <Route path="/completed-expedition-group/:group" component={CompletedExpeditionGroup} />
             <Route path="/field-book" component={FieldBook} />
