+++ src/pages/completed-expedition-group.jsx
@@ -2,14 +2,21 @@ import React, { PropTypes } from 'react';
 import { connect } from 'react-redux';
 import { Hero } from 'components/hero';
 import { FatFooter } from 'components/fat-footer';
-import CompletedExpeditionTiles from 'components/completed-expeditions/tiles';
 import { expeditionsInGroup } from 'helpers/expeditions';
 import { getExpeditionGroup } from 'helpers/expedition-groups';
+import CompletedExpeditionTile from 'components/completed-expedition-tile';
+import { findExpedition } from 'helpers/expeditions';
+import dateformat from 'dateformat';
 
 const CompletedExpeditionGroup = ({ params, inactiveWorkflows }) => {
   const { group } = params;
   const expeditionGroup = getExpeditionGroup(group);
-  const expeditions = expeditionsInGroup(group, inactiveWorkflows);
+  const expeditions = expeditionsInGroup(group, inactiveWorkflows).map(workflow => {
+    const expedition = findExpedition(workflow.display_name);
+    expedition.completed = expedition.completed_at ||
+      dateformat(workflow.finished_at, 'mmmm d yyyy');
+    return expedition;
+  }).sort((a, b) => new Date(b.completed) - new Date(a.completed));
   return (
     <div className="completed-expedition-group">
       <Hero
@@ -18,7 +25,11 @@ const CompletedExpeditionGroup = ({ params, inactiveWorkflows }) => {
       />
       <div className="content">
         { React.createElement(expeditionGroup.icon) }
-        <CompletedExpeditionTiles inactiveWorkflows={expeditions} />
+        <div className="tiles">
+          {expeditions.map((expedition, i) =>
+            <CompletedExpeditionTile expedition={expedition} key={i} />
+          )}
+        </div>
       </div>
       <FatFooter />
     </div>
@@ -32,8 +43,8 @@ CompletedExpeditionGroup.propTypes = {
 
 function mapStateToProps(state) {
   return {
-    // inactiveWorkflows: state.workflows.activeWorkflows,
-    inactiveWorkflows: state.workflows.inactiveWorkflows,
+    inactiveWorkflows: state.workflows.activeWorkflows,
+    // inactiveWorkflows: state.workflows.inactiveWorkflows,
   };
 }
